requirements.txt
Django==4.1.5
djangorestframework==3.14.0
asgiref==3.6.0
certifi==2019.3.9
chardet==3.0.4
charset-normalizer==3.0.1
django-countries==5.3.3
django-crispy-forms==1.7.2
docopt==0.6.2
idna==3.4
pep8==1.7.1
Pillow==9.4.0
pycodestyle==2.5.0
python-decouple==3.1
pytz==2018.5
requests==2.28.2
setuptools==65.5.0
sqlparse==0.2.4
tzdata==2022.7
urllib3==1.24.2
yarg==0.1.9
